requirements.txt
# Core dependencies
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Pretrained models
paddleocr>=2.7.0
timm>=0.9.0
torch>=2.2.0
torchvision>=0.17.0

# ONNX Runtime (choose one based on hardware)
# onnxruntime>=1.16.0  # CPU only
onnxruntime-directml>=1.16.0  # Windows GPU via DirectML

# ML and ranking
lightgbm>=4.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
